Add files via upload
As a failsafe, we made a small modification (a continue statement) to the round_4.py code that skips the buysing and sell procedures for the products. In other words, this code should be submittable to the IMC Dashboard, however it won't do any trading, so this code should end up with 0 overall profit.