Update round_4.py
For context, I was working with round_4_tidied_up.py first, so I was trying to update round_4.py to have the same functionality and result as round_4_tidied_up.py, sort of as a check to make sure our round_4_tidied_up.py code is equivalent to round_4.py; the goal of round_4_tidied_up.py was to have the same result as round_4.py, but to have code that is more organized and hopefully easier to work with. So far, however, round_4_tidied_up.py is getting a higher profit in the IMC Dashboard than round_4.py; I am currently not sure as to why this is happening.