Update round_4_experimental.py
Attempted to add functionality and calculations to calculate the threshold to buy and sell MAGNIFICENT_MACARONS

Essentially, the conversionObservations class provides us with a bunch of information that might influence the MAGNIFICENT_MACARONS price (for now, I will assume that all of them have an effect); these factors include tariff prices, sugar prices, and sunlight. the conversionObservations class also has the bid and ask price of a MAGNIFICENT_MACARONS item, which I am currently not too sure what that is meant to represent (I guess the most recent or a current MAGNIFICENT_MACARONS item?)

From this, we used the factors in the conversionObservations class, stored them in a list history for later, and in subsequent iterations used the history alongside the current conversionObservations class values to get normalized values for each of the factors and summed them up with weights (so far, all of the factors have a weight of 0.1, except sunlight which has a weight of -0.4, as IMC mentioned that sunlight specifically has an impact on the prices of MAGNIFICENT_MACARONS and sugar). Finally, we took the sell and buy order histories of the MAGNIFICENT_MACARONS we got from trading, averaged these histories out, took the mid value of these two averages, and multiplied that by the normalized sum from the conversionObservations class factors; that result is currently used as our threshold to buy and sell MAGNIFICENT_MACARONS

So far, this code is giving us the most profit. However, I worry that this might be an overfitted code.